Building with query_agent
