--- src/cpp/link_creation_agent/remote_iterator.cc
@@ -1,10 +1,0 @@
-#include "remote_iterator.h"
-
-using namespace query_node;
-
-RemoteIterator::RemoteIterator(const string& local_id) : local_id(local_id) {
-    remote_iterator_node = new QueryNodeServer(local_id);
-}
-
-RemoteIterator::~RemoteIterator() {}
-
