+++ src/cpp/link_creation_agent/service.h
@@ -6,7 +6,8 @@
 #include <functional>
 #include <atomic>
 #include "thread_pool.h"
-#include "remote_iterator.h"
+// #include "remote_iterator.h"
+#include "RemoteIterator.h"
 #include "das_server_node.h"
 #include "link.h"
 #include <map>
@@ -14,7 +15,7 @@
 
 using namespace das;
 using namespace query_node;
-
+using namespace query_element;
 namespace link_creation_agent
 {
     class LinkCreationService
